Support "for", "class", "http-equiv", and "accept-charset"
This commit updates the HTML property config and UnknownPropertyHook
developer warnings such that is no longer required to specify
"className" instead of "class", "htmlFor" instead of "for", and so
on.

Both forms are supported. When both are specified within props, React
provides the warning:

"className and class were listed as properties on <div />. Both write
to the same attribute; use one or the other."