DOM Fixtures: Eject CRA. Add new test fixture creator page.
Before this commit, we relied on `window.React` to provide different
versions of React. This doesn't work for npm modules.

To do this, I setup an `externals` configuration in Webpack to specify
that `React`, `ReactDOM`, and `PropTypes` should come from the
window. create-react-app doesn't allow configuration of webpack, so
I've ejected the build and simplied things where possible.