ReactDOM.activeUpdates
Updates scheduled inside of activeUpdates are flushed by the end of
the current tick. In async mode, this means updates are given sync
priority.

In sync mode, updates already have sync priority, so the effect of
activeUpdates is to batch the updates together and flush them at the
end of the batch.

Unlike unstable_batchedUpdates, activeUpdates flushes at the end even if
it's nested inside another batch.