Added 'componentStackFrames' property to error info object
We pass a 'componentStack' (string) property to the injected error dialog and the 'componentDidCatch' lifecycle method. This string is pre-formatted string to match the error.stack string for logging convenience. However this approach requires parsing in order to do anything beyond basic logging.

This PR adds a new 'componentStackFrames' property to both of the above methods. It is loosely based on the TC39 error stack strawman.

In addition to adding basic tests for this property, I've also updated our fixtures to expose it.