Revert "Use the virtual target for change events to avoid restoring controlled state on the real target (#10444)"
This reverts the meaningful (src, non-test) part of commit 3bc64327f01a2c9fd1529934ebf4a484c55d7e94 since we've reverted the commit it depended on in 18083a8a7305483e219592914ef52729907a4368. I don't fully understand the negative implications of leaving this unreverted but it sounds like consensus is that it's safer to revert.

I left the new fixture and verified it works correctly in Chrome 62 Mac, as well as the jest tests passing.