Fix 404ing links in docs sidebar
**what is the change?:** +  **why make this change?:**
Our `slugify` helper was adding two dots to the start of each url in the
sidebar. Maybe for earlier Gatsby this worked, but it was 404ing to
click on a link to "../docs/introducing-jsx.html". This is not related
to the previous commit updating the Gatsby version.

**test plan:**
Manual testing

**issue:**
Checklist item
> "URL generation for eg Contributing and Docs or Community is janky.
> (And started failing after a recent Gatsby upgrade.) Probably to move
> to absolute paths."
https://github.com/bvaughn/react/wiki/Gatsby-check-list