Move installation special case to main router
**what is the change?:**
Instead of adding a fork in the logic in the 'docs' template, we add a
special case in the router itself. That way we branch at a place that
makes sense in the code.

**why make this change?:**
To keep the code organized and more maintainable.

**test plan:**
Manual testing.

**issue:**
Checklist item on gatsby docs update wiki;
https://github.com/bvaughn/react/wiki/Gatsby-check-list
misc > Properly handle tabbed content on /docs/installation.html