Add test for warning for camelCased unknown props
**what is the change?:**
When we render unknown props, they get converted to lower-case.

This may be unexpected for people, or break what they are expecting to
happen. Let's warn in this case and ask them to explicitly pass the
lower-cased attribute name.

**why make this change?:**
To avoid corner case buggy results for users.

**test plan:**
NOTE: ~~~It does NOT pass right now. This is a known issue.
Should we change and just expect a warning, but allow the attribute
value to be set?

`yarn run test src/renderers/dom/shared/__tests__/ReactDOMAttribute-test.js`

**issue:** https://github.com/facebook/react/issues/10399