Remove outdated copy about "non-JSX" view of code playground
**what is the change?:**
Remove a paragraph on the home page that says to "Try clicking on
"Compiled JS" to see the raw JavaScript code", but we don't have the
"Compiled JS" button any more and won't add it for the initial release
of this docs update.

**why make this change?:**
So the copy is accurate.

**test plan:**
Manual testing
(Flarnie will insert a screenshot into the PR)

**issue:**
wiki checklist item on https://github.com/bvaughn/react/wiki/Gatsby-check-list
misc > Change copy to remove reference to "non-JSX" view of interactive code samples.