[Gatsby Docs Update] Stop using 'window' in hacky 'installation' page js
**what is the change?:**
Check for existence of 'window' in one case, avoid using it in another.

**why make this change?:**
When Gatsby executes some of our JS in the node environment, there won't
be a 'window' and we don't want to run into errors there.

This was not breaking but let's be on the safe side.

NOTE that this JS is full of hacks to make it compatible with the old
markdown file and we will completely fix it up in the near future. !

**test plan:**
Manual testing - flarnie will add screenshots.

**issue:**
following up on feedback from https://github.com/facebook/react/pull/10593/commits