[Gatsby docs update] minor fixes (#10642)
* Remove outdated copy about "non-JSX" view of code playground

**what is the change?:**
Remove a paragraph on the home page that says to "Try clicking on
"Compiled JS" to see the raw JavaScript code", but we don't have the
"Compiled JS" button any more and won't add it for the initial release
of this docs update.

**why make this change?:**
So the copy is accurate.

**test plan:**
Manual testing
(Flarnie will insert a screenshot into the PR)

**issue:**
wiki checklist item on https://github.com/bvaughn/react/wiki/Gatsby-check-list
misc > Change copy to remove reference to "non-JSX" view of interactive code samples.

* [Gatsby Docs Update] Stop using 'window' in hacky 'installation' page js

**what is the change?:**
Check for existence of 'window' in one case, avoid using it in another.

**why make this change?:**
When Gatsby executes some of our JS in the node environment, there won't
be a 'window' and we don't want to run into errors there.

This was not breaking but let's be on the safe side.

NOTE that this JS is full of hacks to make it compatible with the old
markdown file and we will completely fix it up in the near future. !

**test plan:**
Manual testing - flarnie will add screenshots.

**issue:**
following up on feedback from https://github.com/facebook/react/pull/10593/commits

* Add link to babel repl

* remove space from link on homepage

* ran prettier