Add urlRoot to `og:url` tag content
**what is the change?:**
See title

**why make this change?:**
We want the `og:url` to have the full url for this page, similar to the
old docs. Unfortunately we can't just grab the current location when
using SSR, so we hard-code the root and add the current pathname, which
we do have access to on the server.

**test plan:**
Manual testing - Flarnie will add screenshots to the PR

**issue:**
Checklist item for React docs overhaul;
Misc > Make sure helmet title and metadata is working correctly