[Gatsby Docs Update] Fix meta tags and ssr docs build (#10705)
* Use `markdownRemark.fields.path` instead of `location.path`

**what is the change?:**
See the title. We updated the graphQL query to request this info for all
pages, and then use it.

Also we had to change the syntax in `gatsby-node.js` because for some
unclear reason, the `fields` came in as `undefined` when using that
shortcut syntax.

**why make this change?:**
- `location` is undefined during the SSR phase, and we want the SSR
  build to succeed.

**test plan:**
Manual testing -
- Ran `yarn build` (!!!) and it worked
- Ran `yarn dev` and manually inspected the `head` element contents.
  Flarnie will insert screenshots.

**issue:**
checklist on the docs site wiki;
https://github.com/bvaughn/react/wiki/Gatsby-check-list
Misc > Meta/Title tags

* remove unused var