Show correct section title based on location
**what is the change?:**
We skim the items in the 'defaultActiveSection' to find the currently
active one, and display it's title in the sticky responsive bottom
navbar.

Before I had just plugged in the section title as a placeholder.

**why make this change?:**
It's better this way. One step closer to Joe's shiny new design. :)

**test plan:**
Manually tested (Flarnie will insert a screenshot)

**issue:**
Sidebar items on this checklist;
https://github.com/bvaughn/react/wiki/Gatsby-check-list