[Gatsby Docs Update] Sidebar + article layout updates (#10735)
* Rename <StickySidebar> to <StickyResponsiveSidebar>

* [Gatsby docs update] Initial version of new sidebar

**what is the change?:**
- Adds media queries and JS to make sidebar collapse/expand from a
  sticky bottom bar on small screen sizes
- Many TODOs still open! This is just an initial version.

**why make this change?:**
Moving closer to Joe's fantastic design. :)

**test plan:**
Manual testing - Flarnie will insert a gif

**issue:**
Various items under the `Sidebar Component` on this checklist;
https://github.com/bvaughn/react/wiki/Gatsby-check-list

* Show correct section title based on location

**what is the change?:**
We skim the items in the 'defaultActiveSection' to find the currently
active one, and display it's title in the sticky responsive bottom
navbar.

Before I had just plugged in the section title as a placeholder.

**why make this change?:**
It's better this way. One step closer to Joe's shiny new design. :)

**test plan:**
Manually tested (Flarnie will insert a screenshot)

**issue:**
Sidebar items on this checklist;
https://github.com/bvaughn/react/wiki/Gatsby-check-list