[Gatsby Docs Update] Tutorial: Clicking an anchor tag should auto-close mobile nav (#10844)
* [Gatsby Docs Update] Tutorial: Clicking an anchor tag should auto-close mobile nav

**what is the change?:**
We create a prop to close the mobile nav and then pass it all the way
down through to the links for the tutorial page -
```
<StickyResponsiveSidebar> -> sends down handler to close mobile nav
V
<Sidebar>
V
<Section>
V
<GatsbyLink> -> receives onClick to close mobile nav
```

**why make this change?:**
So that the menu doesn't stay open after people click a link :)

We considered listening to 'onhashchange' and ran into two problems;
- React doesn't support listening to `onhashchange` on DOMComponents,
  and I'm not sure that would work anyway
- When listening to `window.onhashchange` it didn't seem to fire
- Better to pass something down than to add another global event
  listener imo

**test plan:**
Manual testing
AND I ran `yarn build`

**issue:**
Checklist item on this wiki;
https://github.com/bvaughn/react/wiki/Gatsby-check-list
Sidebar Component > Tutorial: Clicking an anchor tag should auto-close mobile nav

* ran prettier

* Fix label on tutorial page

**what is the change?:**
Skim through the items in the 'defaultActiveSection' and find the one
matching the current location hash.

**why make this change?:**
On the tutorial page it feels weird if the navbar section doesn't match
the section you are on.

**test plan:**
Manual testing + `yarn build`

* fix bad resolution of rebase conflict

* ran prettier

* Tutorial links now update active highlight on-click

* Renamed some Section props for clarity

* Refactored createLink() to use named params