Fix issue where we look for 'namespaceURI' on possibly null parentNode
**what is the change?:**
We have come across a use case where React renders into a comment node
with no parent, and this code assumes that there is always a parent node
for a comment node.

It looks like we should always have a 'Container' type coming in as the
`rootContainerInstance` so it should be safe to check for the
`namespaceURI` there if there is no `parentNode`.

**why make this change?:**
Because we are getting bug reports of this spot throwing due to the
`parentNode` being null. Details tbd, but I imagine it could be a case
where the parentNode was removed by non-React code.

**test plan:**
While this is getting reviewed and we try to get it landed, I'd like to
come up with a test by figuring out how this condition is being reached.

**issue:**
See internal task https://our.intern.facebook.com/intern/tasks/?t=22198152